@@ -8,7 +8,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
-import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.TextMessage;
@@ -40,7 +39,7 @@ public class DashboardHandler extends TextWebSocketHandler {
     private final double GB = 1024 * 1024 * 1024; // 1GB
 
     @Override
-    public void afterConnectionEstablished(@NonNull WebSocketSession session) throws Exception {
+    public void afterConnectionEstablished(WebSocketSession session) throws Exception {
         // 누군가 접속하면 명단에 추가
         sessions.add(session);
         System.out.println("새로운 모니터링 접속자: " + session.getId());
@@ -50,14 +49,14 @@ public class DashboardHandler extends TextWebSocketHandler {
     }
 
     @Override
-    public void afterConnectionClosed(@NonNull WebSocketSession session, @NonNull CloseStatus status) throws Exception {
+    public void afterConnectionClosed(WebSocketSession session, CloseStatus status) throws Exception {
         // 나가면 명단에서 제거
         sessions.remove(session);
         System.out.println("모니터링 접속 해제: " + session.getId());
     }
 
     // 모든 사람에게 메시지(JSON)를 쏘는 기능
-    public void broadcast(@NonNull String message) {
+    public void broadcast(String message) {
         for (WebSocketSession session : sessions) {
             if (session.isOpen()) {
                 synchronized (session) { // 세션별 동기화
